# Memory allocation in C