--- a/devito/__init__.py
+++ b/devito/__init__.py
@@ -3,14 +3,14 @@ import gc
 
 from sympy.core import cache
 
-from devito.interfaces import *  # noqa
-from devito.interfaces import _SymbolCache
 from devito.operator import *  # noqa
 from devito.finite_difference import *  # noqa
 from devito.iteration import *  # noqa
 from devito.dimension import *  # noqa
 from devito.expression import *  # noqa
 from devito.stencilkernel import *  # noqa
+from devito.interfaces import *  # noqa
+from devito.interfaces import _SymbolCache
 
 
 def clear_cache():