--- a/devito/codeprinter.py
+++ b/devito/codeprinter.py
@@ -1,5 +1,4 @@
 from mpmath.libmp import prec_to_dps, to_str
-
 from sympy import Eq
 from sympy.printing.ccode import CCodePrinter