--- a/devito/interfaces.py
+++ b/devito/interfaces.py
@@ -1,16 +1,15 @@
 import weakref
 
 import numpy as np
-from sympy import Function, IndexedBase, as_finite_diff
+from sympy import Function, IndexedBase, as_finite_diff, symbols
 from sympy.abc import h, p, s
 
 from devito.dimension import t, x, y, z
 from devito.finite_difference import (cross_derivative, first_derivative, left,
                                       right)
-from devito.logger import error, logger
+from devito.logger import debug, error
 from devito.memmap_manager import MemmapManager
-from devito.memory import malloc_aligned, free, first_touch
-
+from devito.memory import first_touch, free, malloc_aligned
 
 __all__ = ['DenseData', 'TimeData', 'PointData']
 
@@ -45,7 +44,7 @@ class CachedSymbol(object):
     def _cached_init(self):
         """Initialise symbolic object with a cached object state"""
         original = _SymbolCache[self.__class__]
-        self.__dict__ = original().__dict__.copy()
+        self.__dict__ = original().__dict__
 
 
 class SymbolicData(Function, CachedSymbol):
@@ -113,7 +112,7 @@ class DenseData(SymbolicData):
 
     Note: :class:`DenseData` objects are assumed to be constant in time and
     therefore do not support time derivatives. Use :class:`TimeData` for
-    time-varying griad data.
+    time-varying grid data.
     """
     def __init__(self, *args, **kwargs):
         if not self._cached():
@@ -131,7 +130,7 @@ class DenseData(SymbolicData):
             self.initializer = initializer
             self._data = kwargs.get('_data', None)
             MemmapManager.setup(self, *args, **kwargs)
-            self.memory_allocated = False
+            self.internal_pointer = None
 
     @classmethod
     def _indices(cls, **kwargs):
@@ -153,7 +152,10 @@ class DenseData(SymbolicData):
                 raise ValueError("Unknown symbol dimensions or shape")
             _indices = [x, y, z]
             shape = kwargs.get('shape')
-            dimensions = _indices[:len(shape)]
+            if len(shape) <= 3:
+                dimensions = _indices[:len(shape)]
+            else:
+                dimensions = [symbols("x%d" % i) for i in range(1, len(shape) + 1)]
         return dimensions
 
     @property
@@ -172,7 +174,6 @@ class DenseData(SymbolicData):
         :return: Index corrosponding to the indices
         """
         indices = [a.subs({h: 1, s: 1}) for a in self.args]
-
         return self.indexed[indices]
 
     def _allocate_memory(self):
@@ -184,25 +185,15 @@ class DenseData(SymbolicData):
             self._data = np.memmap(filename=self.f, dtype=self.dtype, mode='w+',
                                    shape=self.shape, order='C')
         else:
-            self._data = self._data, self.internal_pointer = malloc_aligned(
+            self._data, self.internal_pointer = malloc_aligned(
                 self.shape, dtype=self.dtype)
-            logger.info("Allocating memory for %s (%s)" % (self.name, str(self.shape)))
+            debug("Allocating memory for %s (%s)" % (self.name, str(self.shape)))
             first_touch(self)
-            self.memory_allocated = True
 
     def __del__(self):
-        if self.memory_allocated:
+        if self.internal_pointer is not None:
             free(self.internal_pointer)
-            self.memory_allocated = False
-
-    @property
-    def c_element(self):
-        """String representing a single element of this array as it would be
-           referenced in C code
-
-        :returns: String that represents C code to access a single element
-        """
-        return self.name+"["+"][".join(["i"+str(n+1) for n in range(len(self.shape))])+"]"
+            self.internal_pointer = None
 
     @property
     def data(self):
@@ -394,15 +385,6 @@ class TimeData(DenseData):
             self._data = self._data[self.time_order:, :, :]
 
     @property
-    def c_element(self):
-        """String representing a single element of this array as it would be
-           referenced in C code
-
-        :returns: String that represents C code to access a single element
-        """
-        return self.name+"[i"+str(len(self.shape))+"]["+"][".join(["i"+str(n+1) for n in range(len(self.shape)-1)])+"]"
-
-    @property
     def dim(self):
         """Returns the spatial dimension of the data object"""
         return len(self.shape[1:])
@@ -457,7 +439,7 @@ class CoordinateData(SymbolicData):
             self.dtype = kwargs.get('dtype', np.float32)
             self.data, self.internal_pointer = malloc_aligned(
                 self.shape, dtype=self.dtype)
-            self.memory_allocated = True
+            first_touch(self)
 
     def __new__(cls, *args, **kwargs):
         ndim = kwargs.get('ndim')
@@ -472,7 +454,7 @@ class CoordinateData(SymbolicData):
         :param shape: Shape of the spatial data
         :return: indices used for axis.
         """
-        _indices = [p]
+        _indices = [p, s]
         return _indices
 
     @property
@@ -481,18 +463,9 @@ class CoordinateData(SymbolicData):
         return IndexedData(self.name, shape=self.shape, function=self)
 
     def __del__(self):
-        if self.memory_allocated:
-            free(self.data, self.internal_pointer)
-            self.memory_allocated = False
-
-    @property
-    def c_element(self):
-        """String representing a single element of this array as it would be
-           referenced in C code
-
-        :returns String that represents C code to access a single element
-        """
-        return self.name+"["+"][".join(["i"+str(n+1) for n in range(len(self.shape))])+"]"
+        if self.internal_pointer is not None:
+            free(self.internal_pointer)
+            self.internal_pointer = None
 
 
 class PointData(DenseData):
@@ -536,9 +509,8 @@ class PointData(DenseData):
         :param shape: Shape of the spatial data
         :return: indices used for axis.
         """
-        _indices = [t, x, y, z]
-        shape = kwargs.get('shape')
-        return _indices[:len(shape) + 1]
+        _indices = [t, p]
+        return _indices
 
 
 class IndexedData(IndexedBase):