--- a/devito/memory.py
+++ b/devito/memory.py
@@ -1,17 +1,18 @@
 from __future__ import absolute_import
 
 import ctypes
+import operator
 from ctypes.util import find_library
-from operator import mul
 
 import numpy as np
 from sympy import Eq
 from sympy.abc import p
 
-import devito as interfaces
 from devito.iteration import Iteration
-from devito.propagator import Propagator
+from devito.logger import error
 from devito.tools import convert_dtype_to_ctype
+
+mul = operator.mul
 
 
 def malloc_aligned(shape, alignment=None, dtype=np.float32):
@@ -27,20 +28,26 @@ def malloc_aligned(shape, alignment=None, dtype=np.float32):
     needed only for the call to free.
     """
     libc = ctypes.CDLL(find_library('c'))
-    data_pointer = ctypes.c_void_p()
+    data_pointer = ctypes.cast(ctypes.c_void_p(), ctypes.POINTER(ctypes.c_float))
     arraysize = int(reduce(mul, shape))
     ctype = convert_dtype_to_ctype(dtype)
     if alignment is None:
         alignment = libc.getpagesize()
-    libc.posix_memalign(
+
+    ret = libc.posix_memalign(
         ctypes.byref(data_pointer),
         alignment,
-        arraysize * ctypes.sizeof(ctype)
+        ctypes.c_ulong(arraysize * ctypes.sizeof(ctype))
     )
+    if not ret == 0:
+        error("Unable to allocate memory for shape %s", str(shape))
+        return None
+
     data_pointer = ctypes.cast(
         data_pointer,
         np.ctypeslib.ndpointer(dtype=dtype, shape=shape)
     )
+
     pointer = np.ctypeslib.as_array(data_pointer, shape=shape)
     return (pointer, data_pointer)
 
@@ -57,15 +64,18 @@ def first_touch(array):
     """Uses the Propagator low-level API to initialize the given array(in Devito types)
     in the same pattern that would later be used to access it.
     """
+    from devito.propagator import Propagator
+    from devito.interfaces import TimeData, PointData
+
     exp_init = [Eq(array.indexed[array.indices], 0)]
     it_init = []
-    if isinstance(array, interfaces.TimeData):
+    if isinstance(array, TimeData):
         shape = array.shape
         time_steps = shape[0]
         shape = shape[1:]
         space_dims = array.indices[1:]
     else:
-        if isinstance(array, interfaces.PointData):
+        if isinstance(array, PointData):
             it_init = [Iteration(exp_init, index=p, limits=array.shape[1])]
             exp_init = []
             time_steps = array.shape[0]