--- a/devito/tools.py
+++ b/devito/tools.py
@@ -1,4 +1,5 @@
 import ctypes
+
 import numpy as np
 from sympy import symbols