--- a/tests/test_value_params.py
+++ b/tests/test_value_params.py
@@ -1,12 +1,12 @@
 import numpy as np
+from cgen import Assign
 
-import devito.cgen_wrapper as cgen
 from devito.propagator import Propagator
 
 
 def test_value_param():
     data = np.arange(6, dtype=np.float64).reshape((3, 2))
-    kernel = cgen.Assign("output_grid[i2][i1]", "input_grid[i2][i1] + offset")
+    kernel = Assign("output_grid[i2][i1]", "input_grid[i2][i1] + offset")
     propagator = Propagator("process", 3, (2,), [])
     propagator.add_param("input_grid", data.shape, data.dtype)
     propagator.add_param("output_grid", data.shape, data.dtype)