Memory allocation using malloc_aligned. Also implements first_touch.