@navjotk 2016-10-06
I have now addressed all the comments so far. This version is also passing all tests. I believe it should be good to go.