@mlange05 2016-10-06
I'd really like to see at least a simple preliminary _before-and-after_ comparison on a system with multiple NUMA domains before merging this. I just really want to be sure we are actually achieving what we think we're achieving and that we're not impacting the performance of our current benchmarks.