@navjotk 2016-10-06
Since @FabioLuporini is leading the benchmarking cause, he is aware of a lot more "best-practices" and "recipes" than me. I was hoping it would be trivial for him to compare performance of this branch compared to master. Can you help, @FabioLuporini ?