@navjotk 2016-10-06
Tests on bigger domains show that this entire approach fails for larger domains (which are when this approach is required). 
More tests required before this is merged.