@FabioLuporini 2016-10-07
Experimentation is not always about just running tests, but also making sense of them :) so I prefer to instruct you on how to run tests, also because I spent a considerable amount of time on reproducibility.

First thing: I'm gonna merge #136 today, and any tests that you run should be rebased on the new master. This is because that branch fixes several performance measurement issues.

Secondly, to facilitate experimentation, you can temporarily cherry-pick on your branch this commit ``4b0f4c7``, which gives you a ``launcher.sh`` script that you can invoke (``examples`` folder) to run various types of experiments. If you use this, it becomes easy to reproduce exactly the same experiment. 

Now, to the PR.

Once we have some compelling evidence that we are dealing with NUMA correctly (runtime improvements aren't enough; we need some profiling, or perhaps -- simpler -- using a tool such as ``http://man7.org/linux/man-pages/man8/numastat.8.html``), then we should write tests (showing that we aren't leaking memory, for example) and an example for performance reproducibility. The latter can be done by taking an example, and running it on a NUMA node in default mode and with ``first_touch`` somehow disabled.