@navjotk 2016-10-07
Did I mention this PR cuts the memory footprint of the objects in half?