@mlange05 2016-10-07
I can totally believe that, removing the deep copy looks like a very important fix. Well spotted!