@navjotk 2016-10-11
**TLDR: This branch doubles the Gflops based performance and reduces the memory footprint in half.** 


As of now this branch passes all tests on travis and all manual tests I ran here on larger domains (I have tested domains up to 500x500x500).
Without the NUMA related improvement, this branch already performs better than master in terms of GFLOPS/sec (28 Gflops on master vs 34 on this branch consistently across multiple runs). Master also currently allocates at least twice the required amount of memory because of the implementation of the python based alignment function and there is also an in-memory copy there which takes significant time for large domains. Removing all that, this branch reduces the memory footprint in less than half. The time saved in the in-memory copy is probably being offset by the additional overhead of more JIT operations but since a O(n) operation is being replaced by a O(1) one, we should see a total runtime improvement for larger domains. 

Now, coming to the NUMA related improvements. Running on a single socket using `numactl`, we achieve 34 Gflops on a 300x300x300 domain on a node in Yemoja. Clear proof that first touch is happening would be if we can double the performance by running on both sockets. If we run it without `numactl`, the performance varies greatly between 40-70 - my suspicion was that this is due to threads migrating from one core to another. Rerunning it with `KMP_AFFINITY` set to pin each thread to a particular core explicitly confirmed this suspicion and we achieved 70 GFlops consistently - hence confirming that we are indeed using our NUMA channels optimally. 

However, this performance was reached on a version of this code that was slightly modified from the version on this branch. This is because first touch initialization can not happen unless the initialization function has information about the compiler. Since the initialization function is being called from `DenseData`, our `DenseData` objects need a reference to the compiler object that will later be passed to the `Propagator` used to access it. It makes perfect sense since the compiler defines some aspects of how the `Propagator` will access the data structure and without knowing the access pattern, first touch initialization can not happen. In this case, it is the OpenMP flag but in the future we might be using more compiler specific initializations (e.g. if we start to set affinity internally). I am open to ideas on how to structure this in a way that is not considered "hackish". 

Another thing is that these tests were without blocking. When the `Propagator` that will access the data has blocking enabled, the initialization function needs that information along with the block sizes that the `Propagator` will use. How to get that information inside the initialization function is another thing that I want to hear your "non-hackish" ideas on.