@FabioLuporini 2016-10-12
Allright, some questions:

- When you say "Without the NUMA related improvement, this branch already performs better than master in terms of GFLOPS/sec (28 Gflops on master vs 34 on this branch consistently across multiple runs). " -- do you know why is that the case ?
- as for "halving the memory requirements" -- do you mean that "Python" (u mean NumPy, right?) first allocates a memory region, and then to make it aligned, it performs a second copy? I'm not sure I understand, could you point me to some references? 
- feels like you're mixing thread pinning with NUMA-awarness. Obviously, thread pinning should always happen, even on a single socket. Those 70GFlops/s -- I assume you're on your branch. What if you run the same test, on a dual socket machine, clearly with thread pinning, but from ``master`` ? The achieved flop rate should be > 34 but much less than 70. In which case, we would have proof that something good is happening, but still not enough to claim "we have a perfectly working first touch implementation" (ie, we would have indirect evidence supporting our hypothesis, but probably we need to go deeper with profiling to be 100%). 

The rest looks all very good

Thanks