@navjotk 2016-10-13
On master: 
```
[navjot.kukreja@service7 devito]$ KMP_AFFINITY="granularity=fine,proclist=[0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39],explicit"  python examples/benchmark.py -P acoustic -d 300 300 300 -o
Applying Forward
IntelCompiler: Compiling /tmp/devito-1078/8e13b48868e8b6bb5cb052d76c13c9877168adaf.cpp
Shape: (320 x 320 x 320) - Cache Blocking: None
Time: 4.225890 s (1133 MCells/s)
Stencil: 1.687500 OI, 31.21 GFlops/s (time: 4.223773 s)
```

On this branch: 
```
[navjot.kukreja@service7 devito]$ KMP_AFFINITY="granularity=fine,proclist=[0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39],explicit"  python examples/benchmark.py -P acoustic -d 300 300 300 -o
IntelCompiler: Compiling /tmp/devito-1078/bcf80c70e8d713dcf18ce604381444f5249e081d.cpp
IntelCompiler: Compiling /tmp/devito-1078/7ea5b123bfc59e6b236b4e8d02143a482f53e940.cpp
IntelCompiler: Compiling /tmp/devito-1078/915211830732f5771c942910fb30e2a60406cc48.cpp
Applying Forward
IntelCompiler: Compiling /tmp/devito-1078/becf8f272f6129e85929cf25d453da17a945ee0d.cpp
IntelCompiler: Compiling /tmp/devito-1078/b8ff4d07ab6825a0059c84f2d8fb62bfec4004f2.cpp
IntelCompiler: Compiling /tmp/devito-1078/d2432c5ef46ba35b2a7cd547d348f0618103e6c3.cpp
IntelCompiler: Compiling /tmp/devito-1078/e20a64d85aefde306afdb1c9d658d4c2b1d8d629.cpp
IntelCompiler: Compiling /tmp/devito-1078/b6b8b5c89db0512de3fc12d7096d7a6f8ac392be.cpp
Shape: (320 x 320 x 320) - Cache Blocking: None
Time: 1.903779 s (2516 MCells/s)
Stencil: 1.687500 OI, 69.33 GFlops/s (time: 1.901353 s)
```