@navjotk 2016-10-17
On discussing with @mloubout , it was clear that the effects of autotuning only show up on domains of size > 512x512x512. The following are runs with autotuning and cache blocking on, on both branches:
On master:
```
[navjot.kukreja@service8 devito]$ time KMP_AFFINITY="granularity=fine,proclist=[0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39],explicit"  python examples/benchmark.py -P acoustic -d 512 512 512 -o -a
AutoTuner: Start. Mode: brute force
AutoTuner: Number of block sizes that will be attempted: 3
IntelCompiler: Compiling /tmp/devito-1078/1b85fc83eb77f748319a55bbf439c3c7c40be623.cpp
Stencil: 1.687500 OI, 36.80 GFlops/s (time: 0.331378 s)
Stencil: 1.687500 OI, 41.68 GFlops/s (time: 0.292582 s)
Stencil: 1.687500 OI, 41.29 GFlops/s (time: 0.295369 s)
AutoTuner: Finish.
AutoTuner: Estimated runtime for ForwardOperator and 149 time steps: 0.004037 hours
Applying Forward
AutoTuner: Picked: [6, 6, None]
IntelCompiler: Compiling /tmp/devito-1078/67b6046dd79e33b517b83741341547e51ea7c38a.cpp
Shape: (532 x 532 x 532) - Cache Blocking: [6, 6, None]
Time: 19.508382 s (1137 MCells/s)
Stencil: 1.687500 OI, 31.05 GFlops/s (time: 19.505693 s)

real	21m46.799s
user	10m18.481s
sys	25m22.133s
```

On this branch:
```
[navjot.kukreja@service8 devito]$ time KMP_AFFINITY="granularity=fine,proclist=[0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39],explicit"  python examples/benchmark.py -P acoustic -d 512 512 512 -o -a
IntelCompiler: Compiling /tmp/devito-1078/9bef9362e33099c89a81c8433a8c3db71a3947f6.cpp
IntelCompiler: Compiling /tmp/devito-1078/a7e646699ecddf6764e74a0ab56de05d981c9d9f.cpp
IntelCompiler: Compiling /tmp/devito-1078/ad7779c9e2a14d50176ce11de301c7921a2a4fed.cpp
IntelCompiler: Compiling /tmp/devito-1078/986abc438929c93cb5712ef476c0eaae2167dc4b.cpp
IntelCompiler: Compiling /tmp/devito-1078/6c7de65a680602ddf24e2b01b1c2abe18466158a.cpp
AutoTuner: Start. Mode: brute force
AutoTuner: Number of block sizes that will be attempted: 3
IntelCompiler: Compiling /tmp/devito-1078/defc3fb56d861c5adc4ca27a6114dc6ce5f24cc3.cpp
IntelCompiler: Compiling /tmp/devito-1078/eb53275a50a1054e8a2ece64a63c37b289fbefc4.cpp
IntelCompiler: Compiling /tmp/devito-1078/aae0e4900e129e480149373d44cf431e709371f9.cpp
Stencil: 1.687500 OI, 83.02 GFlops/s (time: 0.146908 s)
Stencil: 1.687500 OI, 87.35 GFlops/s (time: 0.139626 s)
Stencil: 1.687500 OI, 83.40 GFlops/s (time: 0.146238 s)
AutoTuner: Finish.
AutoTuner: Estimated runtime for ForwardOperator and 149 time steps: 0.001927 hours
Applying Forward
AutoTuner: Picked: [6, 6, None]
IntelCompiler: Compiling /tmp/devito-1078/c106798c58ea6b4ef612e9507ed48c5a34e91402.cpp
IntelCompiler: Compiling /tmp/devito-1078/f35a00aee7a39fb37cc3401b70f46903ab065f7a.cpp
IntelCompiler: Compiling /tmp/devito-1078/011d5534b980daa610b191f31b5f0c6f206f02f9.cpp
IntelCompiler: Compiling /tmp/devito-1078/6253df18c9f885331b5579b8d6d4dfa1d7a280f9.cpp
IntelCompiler: Compiling /tmp/devito-1078/26b7520568eda7c3ec7700dea66c876b5a8a1b1f.cpp
Shape: (532 x 532 x 532) - Cache Blocking: [6, 6, None]
Time: 7.054673 s (3144 MCells/s)
Stencil: 1.687500 OI, 85.89 GFlops/s (time: 7.052508 s)

real	1m23.088s
user	9m52.758s
sys	1m37.960s
```

@FabioLuporini I did these tests even though your best justification for them was "it takes 5 minutes" (which wasn't even true - look at the times above. The one on master takes 22 minutes for a single run). For me these tests were no better justified than studying the effects of the latitude and longitude of the server location on performance. Do you want me to try that next? :)

I don't understand what additional information would a roofline provide, when the discussion is about the relative performance between two versions of the same code. System specifics don't matter when both versions are being run on the same machine. So again, I don't understand the relevance of what you are asking.